% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BTMetric.R
\name{BTMetric}
\alias{BTMetric}
\title{A model that selects BT for stops.
Trained against arbitrary threshold.}
\usage{
BTMetric(volume_report, thsd_p, thsd_n)
}
\arguments{
\item{volume_report}{a generated frequency graph stop volume overview (see read_whole_folder_RTW)}

\item{thsd_p}{a double of threshold value for metrics. This can be used in a loop to test common threshold historically.}

\item{thsd_n}{a threshold for the number of packages at a stop}
}
\value{
boxtrucks a boxtrucks of size stations x stations$max_stops that is a list of stops that require a box truck
}
\description{
A model that selects BT for stops.
Trained against arbitrary threshold.
}
\details{
beta
}
\examples{
\dontrun{
holder <- c("PLACE")
holder
}
}
