% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clear_all_and_set.R
\name{clear_all_and_set}
\alias{clear_all_and_set}
\title{Clear Local Objects/ Set Library
This function manually removes all local variables and then sets the working directoy proximal to the portable built for this SWA and the library
it should be used if the library fails to import, or if issues are experienced running read_whole_folder}
\usage{
clear_all_and_set()
}
\description{
Clear Local Objects/ Set Library
This function manually removes all local variables and then sets the working directoy proximal to the portable built for this SWA and the library
it should be used if the library fails to import, or if issues are experienced running read_whole_folder
}
\examples{
\dontrun{clear_all_and_set()
}
}
