% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/employeeStopTimeAVG.R
\name{employeeStopTimeAVG}
\alias{employeeStopTimeAVG}
\title{Average employee stop time
Takes a csv of different employee averages and transforms it into a performance metric
This is an effective way of getting a quick look at average employee effieciency along any given route.}
\usage{
employeeStopTimeAVG()
}
\value{
average employee stop time
}
\description{
Average employee stop time
Takes a csv of different employee averages and transforms it into a performance metric
This is an effective way of getting a quick look at average employee effieciency along any given route.
}
\examples{
\dontrun{
mat.empStopTime <- employeeStopTimeAVG()
mat.empStopTime
}
}
