% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTable.R
\name{getTable}
\alias{getTable}
\title{A getter that generates stop frequency matrix}
\usage{
getTable(stationStr)
}
\arguments{
\item{stationStr}{a single vector string}
}
\value{
a matrix that contains the vector string token in the filename
}
\description{
This function looks for particular character sets in a string pulled from the filename and reads them.
It is used recursively with getTable and read_whole_folder_RTW, for more information, please refer to the documentation for those functions
}
\examples{
\dontrun{DLA7.mat <- getTable("DLA7")
DLA7.mat
}
}
\seealso{
\code{\link[=read_whole_folder_RTW]{read_whole_folder_RTW()}}
}
