% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LALO_var.R
\name{LALO_var}
\alias{LALO_var}
\title{Latitude and Longitude total differential
Takes latitude and longitude data and extracts a differential
Allows user to see the differences between displaced area as an additional metric for route risk analysis.}
\usage{
LALO_var(fileName)
}
\arguments{
\item{fileName}{the name of a file (character vector, length 1) from comp in csv format}
}
\value{
total differential
}
\description{
Latitude and Longitude total differential
Takes latitude and longitude data and extracts a differential
Allows user to see the differences between displaced area as an additional metric for route risk analysis.
}
\examples{
\dontrun{
tot_diff <- LALO_var(data.csv)
tot_diff
}
}
