% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mac_manifest.R
\name{mac_manifest}
\alias{mac_manifest}
\title{Concatenate frequency tables by station
This function looks for particular character sets in a string pulled from the filename and reads them.
It is used recursively with getTable and read_whole_folder_RTW, for more information, please refer to the documentation for those functions}
\usage{
mac_manifest(viewTable = FALSE, writeFile = FALSE, routeOverviews = FALSE)
}
\arguments{
\item{viewTable}{a boolean that will display the output of the generated table, TRUE, concatTables will print the table}

\item{writeFile}{a boolean that will write the file to the data folder included in the SWA portable designed for Amazon.}

\item{routeOverviews}{a boolean that will print a route overview for each station that shows the stops for each SWA route}
}
\value{
dataframe of generated stop frequency by station
}
\description{
Concatenate frequency tables by station
This function looks for particular character sets in a string pulled from the filename and reads them.
It is used recursively with getTable and read_whole_folder_RTW, for more information, please refer to the documentation for those functions
}
\examples{
\dontrun{
dir <- paste0(paste0("C:/Users/", Sys.getenv("USERNAME")), "/Documents/R/WD/RTW")
setwd(dir)
stationNames <- c("DLA7", "DPS1")
weekStopFreq <- mac_manifest(stationNames, TRUE)
}
}
