% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/name_function.R
\name{name_func}
\alias{name_func}
\title{Takes the string "dat" and pastes an index
This function is mostly a DRY private function that helps generate the name of dataframes}
\usage{
name_func(index)
}
\arguments{
\item{index}{A numeric vector of length 1 that takes an index. Intended to be used in loops for generating large batches of objects (like Anonymous, or any other fucking thing that I constantly have to avoid or move around)}
}
\value{
None
}
\description{
Takes the string "dat" and pastes an index
This function is mostly a DRY private function that helps generate the name of dataframes
}
\examples{
num <- c(1,2,3,4,5)
for(i in 1:length(num)){
 	
	assign(name_func(i), i)
}
ls()
}
