% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_excel_allsheets.R
\name{read_excel_allsheets}
\alias{read_excel_allsheets}
\title{Reads a workbook to a multidimensional matrix
This is a function to take a workbook and transform it into a multidimensional matrix. This function was created by Jeromy Anglim.}
\arguments{
\item{filename}{a vector of type string and length 1 that takes the name of a file}

\item{tibble}{a logical argument that uses tibbles instead of dataframes if set to TRUE}
}
\value{
None
}
\description{
Reads a workbook to a multidimensional matrix
This is a function to take a workbook and transform it into a multidimensional matrix. This function was created by Jeromy Anglim.
}
\examples{
\code # set the current working directory to a folder with an XLS file in it
\code # this code is in a block comment because it requires an XLS file
\code # NOTRUN {
\code #	setwd(PATH_TO_FOLDER)
\code # }
\code # read_excel_allshets(filename)
\code # ls()
}
\author{
Jeromy Anglim
}
