% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_whole_folder_RTW.R
\name{read_whole_folder_RTW}
\alias{read_whole_folder_RTW}
\title{Reads Manifests to Frequency Table}
\usage{
read_whole_folder_RTW(files, ext, writeEachStation = FALSE)
}
\arguments{
\item{files}{a vector of type string that contains the station name}

\item{ext}{a vector object of length 1 that contains the object name extension -- this will also be passed to the filename along with a timestamp.}

\item{writeEachStation}{a logical argument that will write individual fields for each station}
}
\value{
None
}
\description{
Takes a file of manifests and transforms it into a a stop frequency table.
}
\examples{
\dontrun{setwd(PATH_TO_FOLDER)
fileNames <- list.files()
read_whole_folder_RTW(fileNames, "ALL")
ls()
}
}
