% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/route.R
\name{route}
\alias{route}
\title{Adresses in route}
\usage{
route(stationName)
}
\arguments{
\item{stationName}{character vector of length 1 for the name of the station}
}
\description{
Route stops
Takes a workbook and extracts the route stops
}
\details{
This function may be ammended with fl
}
\examples{
\dontrun{
route("DPS1")
}
}
