% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search}
\alias{search}
\title{A simple search function}
\usage{
search(term)
}
\arguments{
\item{term}{a word to search all the folders for}
}
\value{
a list of files with the search term (WOW!)
}
\description{
Takes a fucking word and looks for shit
}
\details{
This is totally some obvious thing that someone somehow overlooked at somepoint because they're an asshole.
}
\examples{
\dontrun{
filesINeed <- search("obviously")
}
}
