% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeStampAndMove.R
\name{timeStampAndMove}
\alias{timeStampAndMove}
\title{Time stamp, and then move
takes a dataframe and moves it to the Data folder built for the R portable, with an added name extension, then adds a time stamp.}
\usage{
timeStampAndMove(dataframe, extension)
}
\arguments{
\item{dataframe}{the name of the dataframe}

\item{extension}{a really fun name}
}
\value{
None
}
\description{
Time stamp, and then move
takes a dataframe and moves it to the Data folder built for the R portable, with an added name extension, then adds a time stamp.
}
\examples{
d1 <- c(1:5)
d2 <- c(2:6)
d3 <- c(3:7)
mat <- cbind(d1, d2, d3)
mat <- as.data.frame(mat)
colnames(mat) <- c(1:3)
rownames(mat) <- c(1:5)
timeStampAndMove(mat, "Omyjjiodi")
}
