% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/concatTables.R
\name{concatTables}
\alias{concatTables}
\title{Concatenate frequency tables by station}
\usage{
concatTables(stations, viewTable = FALSE, writeFile = FALSE)
}
\arguments{
\item{stations}{a vector of type string containing the names of stations}

\item{viewTable}{a logical argument that will display the output of the generated table, TRUE, concatTables will print the table}

\item{writeFile}{a logical arugment that will write the file to the data folder included in the SWA portable designed for Amazon.}
}
\value{
dataframe of generated stop frequency by station
}
\description{
This function looks for particular character sets in a string pulled from the filename and reads them.
It is used recursively with getTable and read_whole_folder_RTW, for more information, please refer to the documentation for those functions
}
\examples{
\dontrun{
dir <- paste0(paste0("C:/Users/", Sys.getenv("USERNAME")), "/Documents/R/WD/RTW")
setwd(dir)
stationNames <- c("DLA7", "DPS1")
weekStopFreq <- concatTables(stationNames, TRUE)
}
}
