% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeStampAndMove.R
\name{timeStampAndMove}
\alias{timeStampAndMove}
\title{Time stamp, and then move
takes a dataframe and moves it to the Data folder built for the R portable, with an added name extension, then adds a time stamp.}
\usage{
timeStampAndMove(dataframe, extension)
}
\arguments{
\item{dataframe}{the name of the dataframe}

\item{extension}{a really fun name}
}
\value{
None
}
\description{
Time stamp, and then move
takes a dataframe and moves it to the Data folder built for the R portable, with an added name extension, then adds a time stamp.
}
\examples{
\code d1 <- c(1:5)
\code d2 <- c(2:6)
\code d3 <- c(3:7)
\code mat <- cbind(d1, d2, d3)
\code mat <- as.data.frame(mat)
\code colnames(mat) <- c(1:3)
\code rownames(mat) <- c(1:5)
\code timeStampAndMove(mat, "SchmooooooNozzle")
}
